+++ drawingnote_application/lib/datas/datas.txt
@@ -1,0 +1,1 @@
+상수, 모델 등 데이터 관리
