+++ drawingnote_application/lib/features/features.txt
@@ -1,0 +1,1 @@
+레이아웃 관리
