+++ drawingnote_application/lib/services/bluetoothmanager.dart
@@ -1,0 +1,3 @@
+import 'package:bluetooth_classic/bluetooth_classic.dart';
+
+class Bluetoothmanager {}
