+++ drawingnote_application/lib/services/services.txt
@@ -1,0 +1,1 @@
+http, bluetooth 등 통신 로직
